+++ android/build.gradle
@@ -1,46 +1,35 @@
 import groovy.json.JsonSlurper
 
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.1'
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
-    }
-}
-
-allprojects {
-    repositories {
-        mavenLocal()
-        jcenter()
-        maven {
-            // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
-            url "$rootDir/../node_modules/react-native/android"
-        }
-    }
-}
+apply plugin: 'com.android.library'
 
 def getLibraryName = { ->
     def jsonFile = file('../package.json')
     def parsedJson = new JsonSlurper().parseText(jsonFile.text)
     def vName = parsedJson.version
-    if (project.hasProperty('isSnapshot')) {
-        def snap = project.property('isSnapshot')
-        if (snap == 'true' || snap == 'True') {
-            if (project.hasProperty('branchName')) {
-                if (branchName) {
-                    vName = vName + "-" + branchName
-                }
-            }
-            vName =  vName + "-SNAPSHOT"
-        }
-    }
     println "VersionName is set to $vName"
     return vName
 }
 
-project.ext.set("versionName", getLibraryName())
+android {
+    compileSdkVersion 25
+    buildToolsVersion "25.0.3"
+    defaultConfig {
+        minSdkVersion 16
+        targetSdkVersion 25
+        versionCode 1
+        versionName getLibraryName()
+    }
+}
+
+dependencies {
+    compile 'com.facebook.react:react-native:+'
+    compile 'com.android.support:appcompat-v7:25.3.1'
+
+    testCompile 'junit:junit:4.12'
+
+    // Android JUnit Runner
+    androidTestCompile 'com.android.support.test:runner:0.5'
+    // JUnit4 Rules
+    androidTestCompile 'com.android.support.test:rules:0.5'
+
+}
