--- android/react-native-android-fragment/.gitignore
@@ -1,1 +1,0 @@
-/build
