--- android/settings.gradle
@@ -1,4 +1,0 @@
-rootProject.name = 'ReactNative-Android-Fragment'
-
-include ':react-native-android-fragment', ':sample-app/android:app'
-
