+++ android/src/main/AndroidManifest.xml
