+++ android/src/main/res/values/strings.xml
