--- example/android/app/.gitignore
@@ -1,1 +1,0 @@
-/build
