+++ example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">SampleApp</string>
+    <string name="app_name">React Native Android Fragment</string>
 </resources>
