+++ index.js
@@ -1,6 +1,7 @@
-
-import { NativeModules } from 'react-native';
-
-const { AndroidFragment } = NativeModules;
-
-export default AndroidFragment;
+/**
+ * No-Op:
+ * 
+ * This native library handles linking the code required for use
+ * in a Native Android project, but it currently doesn't require the need to expose
+ * any bridge methods, views, or constants.
+ */
