+++ package.json
@@ -3,6 +3,11 @@
   "name": "react-native-android-fragment",
   "version": "0.46.2",
   "description": "A utility library for facilitating React Native development with Android Fragments.",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/hudl/react-native-android-fragment.git"
+  },
+  "homepage": "https://github.com/hudl/react-native-android-fragment#readme",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -13,9 +18,6 @@
   "author": "hudl",
   "license": "Apache-2.0",
   "peerDependencies": {
-    "react-native": "^0.47.0"
-  },
-  "devDependencies": {
     "react-native": "^0.47.0"
   }
 }
